requirements.txt
flask
openai
pdfkit
python-dotenv
jinja2